+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - WebConsole: Fix metrics values depend on polling period
+- WebConsole: Improve WebConsole throughput and memory graphs (#491)
 
 ### Added